Make tests reusable
This is helpful in case you curate your own mappings somewhere else and still want to have a test case for them being internally right
